(fix) Renamed listeners mixin
