+++ src/elements/GeoSegmentedControl/GeoWarningSegmentedControlItem.vue
@@ -14,7 +14,7 @@
 
 <script>
 import geoSegmentedControlItemMixin, { VARIANTS } from './GeoSegmentedControlItem.constants'
-import customListenersMixin from '@/mixins/customListenersMixin'
+import extendedListenersMixin from '@/mixins/extendedListenersMixin'
 
 /**
  * `GeoWarningSegmentedControlItem` is a component designed to nicely fit as
@@ -26,7 +26,7 @@ export default {
   name: 'GeoWarningSegmentedControlItem',
   status: 'ready',
   release: '20.1.0',
-  mixins: [geoSegmentedControlItemMixin, customListenersMixin],
+  mixins: [geoSegmentedControlItemMixin, extendedListenersMixin],
   helpers: {
     variant: VARIANTS.warning
   }
